requirements.txt
absl-py==0.11.0
amqp==5.0.2
asn1crypto==0.24.0
astunparse==1.6.3
billiard==3.6.3.0
boto3==1.12.26
botocore==1.15.26
cachetools==4.1.1
celery==5.0.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
configparser==4.0.2
cryptography==2.1.4
cycler==0.10.0
docutils==0.15.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gast==0.3.3
gnureadline==8.0.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.6
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
minio==5.0.8
natsort==7.1.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
SecretStorage==2.3.1
six==1.15.0
SQLAlchemy==1.3.20
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.54.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0